backend/requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
requests
python-dotenv
langchain
nltk
transformers
torch